feat: add premium to role user

+++ src/services/product.services.js
@@ -1,21 +1,31 @@
 import { productResponseDto } from "../dto/product-response.dto.js";
 import productsRepository from "../persistence/mongo/repositories/products.repository.js";
 import error from '../errors/customErrors.js';
+import customErrors from "../errors/customErrors.js";
 
 const getAll = async (query, options) => {
     return await productsRepository.getAll(query, options);
 }
 
-const getById = async (id) => {
-    const productData = await productsRepository.getById(id);
-    if (!productData) throw error.notFound(`Product with id ${id} not found!`);
+const getById = async (pid) => {
+    const productData = await productsRepository.getById(pid);
+    if (!productData) throw customErrors.notFound(`Product with id ${pid} not found!`);
     const product = productResponseDto(productData);
+    
     return product;
 }
 
-const create = async (data) => {
-    const product = await productsRepository.create(data);
-    if (!product) throw error.notFound(`There was an error creating the product!`);
+const create = async (data, user) => {
+    let productData = data;
+
+    if (user.role === 'premium') {
+        productData = {
+            ...data,
+            owner: user._id
+        }
+    }
+    const product = await productsRepository.create(productData);
+    if (!productData) throw error.notFound(`There was an error creating the product!`);
     return product;
 }
 
@@ -25,11 +35,14 @@ const update = async (id, data) => {
     return product
 }
 
-const destroy = async (id) => {
-    const product = await productsRepository.deleteOne(id)
-    if (!product) throw error.notFoundError(`Product id ${id} not found`)
-    return product
-}
+const destroy = async (id, user) => {
+    const productData = await productsRepository.getById(id);
+    if (user.role === "premium" && productData.owner !== user._id) {
+        throw customErrors.unAuthorized("Only the owner can delete a product");
+    }
+    const product = await productsRepository.deleteOne(id);
+    return product;
+};
 
 export default {
     getAll,
